Allow interpolation onto VertexOnlyMesh
WIP for allowing trans-mesh interpolation onto a VertexOnlyMesh by
creating a QuadratureElement with a single runtime specified point and
weight. A QuadratureElement is "A set of quadrature points pretending to
be a finite element"

We manually build a double indirection map for the par_loop argument
that takes you from vertex only mesh cells to parent mesh nodes (the
"mesh coordinates" argument).

Note that we had to manually set the par loop argument order so that the
this argument was last in the list since The loopy kernel argument order
is not what it seems from simply printing the kernel. Instead it is
given by `kernel.code.args`.
The result of this is that the "mesh coordinates" argument actually needs
to come after the pre-interpolation coefficient argument.